add shortguid generator

+++ ShortGuid.h
@@ -5,7 +5,8 @@
 
 namespace CATHODE::ShortGuid
 {
-	int __fastcall hShortGuid(DataTypes::ShortGuid* _this, void* _EDX, char* node_name);
-	typedef int(__thiscall* tShortGuid)(DataTypes::ShortGuid*, char*);
+	uintptr_t __fastcall hShortGuid(DataTypes::ShortGuid* _this, void* _EDX, char* node_name);
+	typedef uintptr_t(__thiscall* tShortGuid)(DataTypes::ShortGuid*, char*);
 	inline auto ShortGuid = reinterpret_cast<tShortGuid>(DEVTOOLS_RELATIVE_ADDRESS(0x004bf7f0));
+
 }
